initial load
Working Java and Scala code for the MMA8451 Accelerometer on a
RaspberryPi 3B
Did copy the full set of Pi4J examples into this workspace, but was
unable to get some of them to be ok with my maven resource set.
Retained those which could be made to have no errors.